requirements.txt
flask
requests
python-dotenv
PyPDF2
huggingface-hub>=0.20.3
